create app for duel
