+++ hire_hire/interview/models.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import get_user_model
 from django.db import models
 
-from interview.managers import QuestionManager, DuelQuestionManager, DuelPlayer
+from interview.managers import QuestionManager
 
 User = get_user_model()
 
@@ -76,101 +76,3 @@ class Interview(models.Model):
 
     def __str__(self):
         return f'интервью {self.pk}'
-
-
-class Duel(models.Model):
-    """
-    Объект дуэли.
-    В MVP:
-        - может содержать овнера
-        - считает неверные ответы
-    """
-    owner = models.ForeignKey(
-        User,
-        on_delete=models.CASCADE,
-        null=True,
-        related_name='duels',
-        verbose_name='владец игры',
-    )
-
-    wrong_answers_count = models.IntegerField(
-        'количество неправильных ответов',
-        default=0,
-    )
-
-    class Meta:
-        verbose_name = 'дуэль'
-        verbose_name_plural = 'дуэли'
-
-    def __str__(self):
-        return f'дуэль {self.pk}'
-
-
-class DuelPlayer(models.Model):
-    """
-    Объект счетчика в дуэли.
-    В MVP:
-        - хранит имя игрока
-        - хранит текущий счет
-        - связана с дуэлью
-    """
-
-    name = models.CharField(
-        'имя игрока',
-        max_length=100,
-        default='Игрок',
-    )
-
-    good_answers_count = models.IntegerField(
-        default=0,
-        verbose_name='счетчик',
-    )
-
-    duel = models.ForeignKey(
-        Duel,
-        on_delete=models.CASCADE,
-        related_name='players',
-        verbose_name='дуэль',
-    )
-    objects = DuelPlayer()
-
-    class Meta:
-        verbose_name = 'участник дуэли'
-        verbose_name_plural = 'участники дуэли'
-
-    def __str__(self):
-        return self.name[:15]
-
-
-class DuelQuestion(models.Model):
-    """
-    Объект вопроса в рамках дуэли.
-    В MVP:
-        - хранит вопрос для конкретной дуэли
-        - хранит статус вопроса
-    """
-
-    duel = models.ForeignKey(
-        Duel,
-        on_delete=models.CASCADE,
-        related_name='questions',
-        verbose_name='дуэль',
-    )
-
-    question = models.ForeignKey(
-        Question,
-        on_delete=models.CASCADE,
-        related_name='duels',
-        verbose_name='вопрос',
-    )
-
-    is_answered = models.BooleanField('дан ответ')
-
-    objects = DuelQuestionManager()
-
-    class Meta:
-        verbose_name = 'вопрос в дуэли'
-        verbose_name_plural = 'вопросы в дуэли'
-
-    def __str__(self):
-        return f'вопрос дуэли {self.pk}'
