+++ hire_hire/interview/urls.py
@@ -1,10 +1,6 @@
 from django.urls import path
 
 from interview.views import (
-    DuelFinishView,
-    DuelFlowAnsweredView,
-    DuelFlowQuestionView,
-    DuelSettingsView,
     InterviewFinishView,
     InterviewFlowView,
     InterviewSettingsView,
@@ -15,37 +11,17 @@ app_name = 'interview'
 
 urlpatterns = [
     path(
-        'duel_settings/',
-        DuelSettingsView.as_view(),
-        name='duel_settings',
-    ),
-    path(
-        'duel/<int:duel_id>/finish/',
-        DuelFinishView.as_view(),
-        name='duel_finish',
-    ),
-    path(
-        'duel/<int:duel_id>/answer/',
-        DuelFlowAnsweredView.as_view(),
-        name='duel_answer',
-    ),
-    path(
-        'duel/<int:duel_id>/',
-        DuelFlowQuestionView.as_view(),
-        name='duel',
-    ),
-    path(
         'settings/',
         InterviewSettingsView.as_view(),
         name='settings',
     ),
     path(
-        'interview/<int:interview_id>/',
+        '<int:interview_id>/',
         InterviewFlowView.as_view(),
         name='interview',
     ),
     path(
-        'interview/finish/',
+        'finish/',
         InterviewFinishView.as_view(),
         name='finish',
     ),
