+++ hire_hire/templates/duel/duel-results.html
@@ -22,7 +22,7 @@
       <a href="{{ url('homepage:index') }}" class="duel-result__link page__link page__text">
         На главную
       </a>
-      <a href="{{ url('interview:duel_settings') }}" class="duel-result__link duel-result__link_type_pouring page__link page__text">
+      <a href="{{ url('duel:duel_settings') }}" class="duel-result__link duel-result__link_type_pouring page__link page__text">
         Новое испытание
       </a>
     </div>
