+++ hire_hire/templates/duel/duel-settings.html
@@ -7,7 +7,7 @@
         <h2 class="duel-settings__subtitle sections__subtitle">
             Python
         </h2>
-        <form method="post" action="{{ url('interview:duel_settings') }}"
+        <form method="post" action="{{ url('duel:duel_settings') }}"
               class="duel-settings__form">
             {{ csrf_input }}
             <label class="duel-settings__label page__text">
