+++ hire_hire/templates/homepage/index.html
@@ -44,7 +44,7 @@
     буквально пару предложений. Короткий поясняющий текст. Короткий поясняющий текст, буквально пару предложений.
     Короткий поясняющий текст. Короткий поясняющий текст, буквально пару предложений. Короткий поясняющий текст.
   </p>
-  <a href="{{ url('interview:duel_settings') }}" class="games__link page__link page__text">
+  <a href="{{ url('duel:duel_settings') }}" class="games__link page__link page__text">
     Провести дуэль
   </a>
 </section>
